feat: adiciona logica para reserva, juntamente de validações feitas nos request de todas as entidades

+++ src/main/java/com/excursao/marcinho/service/OnibusService.java
@@ -3,6 +3,7 @@ package com.excursao.marcinho.service;
 import com.excursao.marcinho.dto.request.OnibusRequest;
 import com.excursao.marcinho.dto.response.OnibusResponse;
 import com.excursao.marcinho.entity.Onibus;
+import com.excursao.marcinho.exceptions.notfound.OnibusNotFoundException;
 import com.excursao.marcinho.mapper.OnibusMapper;
 import com.excursao.marcinho.repository.OnibusRepository;
 import lombok.RequiredArgsConstructor;
@@ -27,13 +28,13 @@ public class OnibusService {
 
     public OnibusResponse findByID(Long id){
         Onibus onibus = repository.findById(id)
-                .orElseThrow(() -> new RuntimeException("O id do onibus não é valido!"));
+                .orElseThrow(OnibusNotFoundException::new);
         return mapper.toResponse(onibus);
     }
 
     public OnibusResponse update (Long id, OnibusRequest onibusRequest){
         Onibus onibus = repository.findById(id)
-                .orElseThrow(() -> new RuntimeException("O id do onibus não é valido!"));
+                .orElseThrow(OnibusNotFoundException::new);
 
         mapper.updateFromResponse(onibusRequest, onibus);
         Onibus atualizado = repository.save(onibus);
@@ -42,6 +43,10 @@ public class OnibusService {
     }
 
     public void delete(Long id){
+        if (!repository.existsById(id)){
+            throw new OnibusNotFoundException();
+        }
+
         repository.deleteById(id);
     }
 }
